feat: chunks 25+26 — artifacts recipe, lint gate, docs; build complete
Chunk 25: docs/ARTIFACTS.md is the reproducible recipe for the detector (download,
hash-pinned) and the species/bird heads (build_coco_manifest / build_cub_manifest +
train), plus the synthetic smoke path. No model weights in git — .acmodel added to
.gitignore alongside .pt/.pth; they are large and reproducible.

Chunk 26: the lint gate F23 asked for — a pinned [tool.ruff] ruleset
(E,F,I,B,UP,C4,PLC,RUF) with intentional ignores (B008 typer idiom, PLC0415 lazy
imports, E501/RUF002 prose docstrings) and ruff==0.14.1 in the dev group;
'uv run --frozen ruff check' is clean. Fixed the substantive findings it surfaced
(zip strict=, redundant int() casts, unused letterbox unpack). README rewritten for
the finished tool with a quick start; docs/test-report.md and PROGRESS.md capture the
green run; impl-status.json is complete=true.

Final state: all 26 chunks done, e2e suite 109 passed (100 fast + 9 slow) / 0 failed
/ 0 xfail, ruff clean. Real pipeline verified end to end on real COCO/MegaDetector
data. Three bugs found and fixed by the tests during the build (dedup, backbone head
load, reclassify vs human labels).

+++ src/animal_classifier/training/dataset.py
@@ -43,7 +43,7 @@ def train_transform(input_size: int) -> transforms.Compose:
 def eval_transform(input_size: int) -> transforms.Compose:
     return transforms.Compose(
         [
-            transforms.Resize(int(round(input_size * 1.14))),
+            transforms.Resize(round(input_size * 1.14)),
             transforms.CenterCrop(input_size),
             transforms.ToTensor(),
             transforms.Normalize(IMAGENET_MEAN, IMAGENET_STD),
